+++ Sources/KeyboardKit/Gestures/ScrollViewGestureButton.swift
@@ -29,7 +29,7 @@ import SwiftUI
  can use the `isPressed` value that is passed to the `label`
  builder, to configure the button view for the pressed state.
  */
-@available(iOS 14.0, macOS 11.0, watchOS 8.0, *)
+@available(iOS 14.0, macOS 11.0, watchOS 7.0, *)
 public struct ScrollViewGestureButton<Label: View>: View {
 
     /**
@@ -128,7 +128,7 @@ public struct ScrollViewGestureButton<Label: View>: View {
     }
 }
 
-@available(iOS 14.0, macOS 11.0, watchOS 8.0, *)
+@available(iOS 14.0, macOS 11.0, watchOS 7.0, *)
 extension ScrollViewGestureButton {
 
     class GestureState: ObservableObject {
@@ -196,7 +196,7 @@ extension ScrollViewGestureButton {
     }
 }
 
-@available(iOS 14.0, macOS 11.0, watchOS 8.0, *)
+@available(iOS 14.0, macOS 11.0, watchOS 7.0, *)
 private extension ScrollViewGestureButton.Style {
 
     func handleIsPressed() {
@@ -219,7 +219,7 @@ private extension ScrollViewGestureButton.Style {
     }
 }
 
-@available(iOS 14.0, macOS 11.0, watchOS 8.0, *)
+@available(iOS 14.0, macOS 11.0, watchOS 7.0, *)
 private extension View {
 
     typealias Action = () -> Void
@@ -291,7 +291,7 @@ private extension GeometryProxy {
     }
 }
 
-@available(iOS 14.0, macOS 11.0, watchOS 8.0, *)
+@available(iOS 14.0, macOS 11.0, watchOS 7.0, *)
 struct ScrollViewGestureButton_Previews: PreviewProvider {
 
     struct Preview: View {